% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{Aggregate}
\alias{Aggregate}
\title{Aggregate}
\usage{
Aggregate(input.files, FUNC, cols = NULL, file.out = NULL)
}
\arguments{
\item{input.files}{List of Parquet files.}

\item{FUNC}{Function to apply on each Parquet table.}

\item{cols}{Columns to subset.}

\item{file.out}{Output Parquet filename.}
}
\value{
The aggregated Parquet tables as a \code{data.table} if
  \code{file.out} is not NULL, otherwise \code{file.out}.
}
\description{
Aggregate several Parquet tables as a single one.
}
