% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{AggregateCommits}
\alias{AggregateCommits}
\title{Aggregate commits}
\usage{
AggregateCommits(gitlog, FUNC, file.out = NULL)
}
\arguments{
\item{gitlog}{Data.frame containing for each Git repository the
Parquet filename of the commit log and the Parquet filename of
the diff.}

\item{FUNC}{Function to apply on each commit table.}

\item{file.out}{Output Parquet filename.}
}
\value{
The aggregated Parquet tables as a \code{data.table} if
  \code{file.out} is not NULL, otherwise \code{file.out}.
}
\description{
Aggregate commits for different repositories as a single table.
}
