% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{AggregateIssues}
\alias{AggregateIssues}
\title{Aggregate issues}
\usage{
AggregateIssues(jiralog, bugzillalog, t, FUNC, file.out = NULL)
}
\arguments{
\item{jiralog}{Data.frame containing all the Jira Parquet
filenames.}

\item{bugzillalog}{Data.frame containing all the Bugzilla Parquet
filenames}

\item{t}{Type of issue log to create (e.g. issue or comment).}

\item{FUNC}{Function to apply on each table.}

\item{file.out}{Output Parquet filename.}
}
\value{
The aggregated Parquet tables as a \code{data.table} if
  \code{file.out} is not NULL, otherwise \code{file.out}.
}
\description{
Aggregate issue data as a single table.
}
