% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bugzilla.R
\name{BugzillaComment}
\alias{BugzillaComment}
\title{Bugzilla Comment}
\usage{
BugzillaComment(comment)
}
\arguments{
\item{comment}{List representing the comment (parsed from JSON).}
}
\value{
A list representing the comment.
}
\description{
Parses a Bugzilla issue comment.
}
