% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bugzilla.R
\name{BugzillaHistoryItems}
\alias{BugzillaHistoryItems}
\title{Bugzilla History Items}
\usage{
BugzillaHistoryItems(items)
}
\arguments{
\item{items}{List of history changes (parsed from JSON).}
}
\value{
A \code{data.table} representing the changes.
}
\description{
Parses a list of Bugzilla history changes
}
