% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bugzilla.R
\name{BugzillaIssue}
\alias{BugzillaIssue}
\title{Bugzilla Issue}
\usage{
BugzillaIssue(issue)
}
\arguments{
\item{issue}{Issue as a list (parsed from JSON).}
}
\value{
List of fields.
}
\description{
Parses a Bugzilla Issue
}
