% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bugzilla.R
\name{BugzillaTime}
\alias{BugzillaTime}
\title{Bugzilla Time}
\usage{
BugzillaTime(table, name)
}
\arguments{
\item{table}{deta.frame containing raw time fields}

\item{name}{Name of the column in which to add the parsed time.}
}
\value{
The table with the time parsed as POSIXct added as a
  column.
}
\description{
Parses Bugzilla raw time fields as POSIXct.
}
