% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identities.R
\name{BuildMergingGraph}
\alias{BuildMergingGraph}
\title{Build merging graph}
\usage{
BuildMergingGraph(people, merging)
}
\arguments{
\item{people}{List of identities as a \code{data.table} object.}

\item{merging}{Merging of people.}
}
\value{
A graph with identities as nodes and merging as edges.
}
\description{
Build an \code{igraph} object based on identities and merging.
}
