% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlp.R
\name{ClassifyTextLines}
\alias{ClassifyTextLines}
\title{Classify text lines}
\usage{
ClassifyTextLines(text, raw)
}
\arguments{
\item{text}{Text.}

\item{raw}{Raw text.}
}
\value{
A Line Type factor.
}
\description{
Classifies text lines.
}
\seealso{
LineType
}
