% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{CleanText}
\alias{CleanText}
\title{Clean Text}
\usage{
CleanText(text)
}
\arguments{
\item{text}{The text.}
}
\value{
The cleaned text.
}
\description{
Clean text by removing URLs, replacing non graphical symbols with
space and trailing white spaces.
}
