% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{CommentTimestamps}
\alias{CommentTimestamps}
\title{Comment Timestamps}
\usage{
CommentTimestamps(comments.input, idmerging)
}
\arguments{
\item{comments.input}{parquet file containing the issue comment log.}

\item{idmerging}{Identity merging data.table.}
}
\value{
The issue comment timestamp log.
}
\description{
Returns list of timestamps for issue comments.
}
