% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identities.R
\name{CommitsIdentities}
\alias{CommitsIdentities}
\title{Commits Identities}
\usage{
CommitsIdentities(commits)
}
\arguments{
\item{commits}{\code{data.table} commit log.}
}
\value{
\code{data.table} with columns source, repo, key,
  displayname, type, name, email and N (number of commits).
}
\description{
Extract different author and committer identities from a set of commits.
}
