% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{CommitTimestamps}
\alias{CommitTimestamps}
\title{Commit Timestamps}
\usage{
CommitTimestamps(commits.input, idmerging)
}
\arguments{
\item{commits.input}{parquet file containing the commit log.}

\item{idmerging}{Identity merging data.table.}
}
\value{
The commit timestamp log.
}
\description{
Returns list of timestamps for commits.
}
