% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{CommitTimeZoneAccuracy}
\alias{CommitTimeZoneAccuracy}
\title{Commit TimeZone Accuracy}
\usage{
CommitTimeZoneAccuracy(log)
}
\arguments{
\item{log}{Commit log as a data.table object.}
}
\value{
The modified commit log with added from.svn and accurate.tz
  columns.
}
\description{
Adds boolean columns (from.svn and accurate.tz) to a data.table of
commits for determining which commits have accurate timezones.
}
