% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{CommitTimeZoneHistory}
\alias{CommitTimeZoneHistory}
\title{Commit Time Zone History}
\usage{
CommitTimeZoneHistory(log.input, idmerging.input, output)
}
\arguments{
\item{log.input}{Parquet file containing the commit log.}

\item{idmerging.input}{Parquet file with identity merging.}

\item{output}{Parquet file where to store the output.}
}
\value{
The output filename (if any) or the data.table object
  otherwise.
}
\description{
Computes timezone history in commits for developer profiles based
on identity merging.
}
