% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identities.R
\name{ExtractEmail}
\alias{ExtractEmail}
\title{Extract Email}
\usage{
ExtractEmail(person)
}
\arguments{
\item{person}{The fields.}
}
\value{
The author fields as a \code{data.table} object with the
  author field, author name amd author email.
}
\description{
Extracts email from Git author fields.
}
