% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{File}
\alias{File}
\title{File}
\usage{
File(
  types,
  repos,
  datadir,
  subdir = ".",
  subname = "",
  ext = "parquet",
  file.func = "none"
)
}
\arguments{
\item{types}{Vector of repository types to keep (e.g. git, jira,
bugzilla).}

\item{repos}{List of repositories as a \code{data.table} as
returned by \code{RawData}.}

\item{datadir}{Directory where the data is stored.}

\item{subdir}{Sub-directory inside \code{datadir}.}

\item{subname}{Added to the name of the data file.}

\item{ext}{Extension of the data file.}

\item{file.func}{If input or output, adds call to drake's
\code{file_in} or \code{file_out}.}
}
\description{
File returns a filenames for a list of repositories.
}
