% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identities.R
\name{FindBots}
\alias{FindBots}
\title{Find bots}
\usage{
FindBots(identities)
}
\arguments{
\item{identities}{List of merged identities.}
}
\value{
The log with an added boolean column is.bot.
}
\description{
Find bots in list of merged identities.
}
