% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{FirstTime}
\alias{FirstTime}
\title{First Time}
\usage{
FirstTime(time, tz)
}
\arguments{
\item{time}{Vector of timestamps.}

\item{tz}{Vector of timezones.}
}
\value{
A POSIXct object.
}
\description{
Returns first timestamp with valid Git timezone (first timestamp
from 2007 and beyond that is not UTC timezone).
}
