% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{GitPlan}
\alias{GitPlan}
\title{Git Plan}
\usage{
GitPlan(repos, datadir)
}
\arguments{
\item{repos}{Repository \code{data.table} object.}

\item{datadir}{Directory where the data is stored.}
}
\value{
A drake plan tibble.
}
\description{
Returns a drake plan for git repositories
}
