% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identities.R
\name{Identities}
\alias{Identities}
\title{Identities}
\usage{
Identities(commits, issues, comments, file.out = NULL)
}
\arguments{
\item{commits}{\code{data.table} with Git commit identities.}

\item{issues}{\code{data.table} with issue identities.}

\item{comments}{\code{data.table} with issue comment
identities.}

\item{file.out}{Parquet output file.}
}
\value{
The \code{data.table} of identities if \code{file.out} is
  NULL, or the output filename otherwise.
}
\description{
Extract identities from commits, issues and issue comments.
}
