% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identities.R
\name{IdentityMerging}
\alias{IdentityMerging}
\title{Identity merging}
\usage{
IdentityMerging(people, file.out = NULL)
}
\arguments{
\item{people}{\code{data.table} with identities to merge.}

\item{file.out}{Output Parquet file.}
}
\value{
Identity merging \code{data.table} object.
}
\description{
Makes identity merging.
}
