% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{IdentityMergingPlan}
\alias{IdentityMergingPlan}
\title{Identity Merging Plan}
\usage{
IdentityMergingPlan(datadir)
}
\arguments{
\item{datadir}{Directory where the data is stored.}
}
\value{
A drake plan tibble.
}
\description{
Returns a drake plan for identity merging
}
