% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identities.R
\name{IssueCommentsIdentities}
\alias{IssueCommentsIdentities}
\title{Issue Comments Identities}
\usage{
IssueCommentsIdentities(comments)
}
\arguments{
\item{comments}{\code{data.table} with issue comment log.}
}
\value{
\code{data.table} with columns source, repo, key,
  displayname, type, name, email and N (number of comments).
}
\description{
Extract different identities from a set of issue comments.
}
