% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{IssuePlan}
\alias{IssuePlan}
\title{Issue Plan}
\usage{
IssuePlan(datadir)
}
\arguments{
\item{datadir}{Directory where the data is stored.}
}
\value{
A drake plan tibble.
}
\description{
Returns a drake plan for aggregating issue and comment logs.
}
