% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identities.R
\name{IssuesIdentities}
\alias{IssuesIdentities}
\title{Issues Identities}
\usage{
IssuesIdentities(issues)
}
\arguments{
\item{issues}{\code{data.table} with issue log.}
}
\value{
\code{data.table} with columns source, repo, key,
  displayname, type, name, email and N (number of commits).
}
\description{
Extract different identities from a set of issues.
}
