% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{IssueTimestamps}
\alias{IssueTimestamps}
\title{Issue Timestamps}
\usage{
IssueTimestamps(issues.input, idmerging)
}
\arguments{
\item{issues.input}{parquet file containing the issue log.}

\item{idmerging}{Identity merging data.table.}
}
\value{
The issue timestamp log.
}
\description{
Returns list of timestamps for issues.
}
