% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jira.R
\name{JiraComment}
\alias{JiraComment}
\title{Jira Comment}
\usage{
JiraComment(comment)
}
\arguments{
\item{comment}{List representing the comment (parsed from JSON).}
}
\value{
A list representing the comment.
}
\description{
Parses a Jira issue comment.
}
