% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jira.R
\name{JiraComments}
\alias{JiraComments}
\title{Jira Comments}
\usage{
JiraComments(issue)
}
\arguments{
\item{issue}{List representing the issue (parsed from JSON).}
}
\value{
A \code{data.table} representing the issue comments.
}
\description{
Parses several issue comments.
}
