% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jira.R
\name{JiraComponents}
\alias{JiraComponents}
\title{Jira Components}
\usage{
JiraComponents(issue)
}
\arguments{
\item{issue}{Issue as a list parsed from JSON.}
}
\value{
A character vector with each component on a separate line.
}
\description{
Gets components from a Jira issue.
}
