% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jira.R
\name{JiraHistory}
\alias{JiraHistory}
\title{Jira History}
\usage{
JiraHistory(issue)
}
\arguments{
\item{issue}{List representing the issue (parsed from JSON).}
}
\value{
A \code{data.table} representing the history of change of the issue.
}
\description{
Parses history from an issue.
}
