% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jira.R
\name{JiraHistoryItems}
\alias{JiraHistoryItems}
\title{Jira History Items}
\usage{
JiraHistoryItems(items)
}
\arguments{
\item{items}{List of history changes (parsed from JSON).}
}
\value{
A \code{data.table} representing the changes.
}
\description{
Parses a list of Jira history changes
}
