% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{JiraPlan}
\alias{JiraPlan}
\title{Jira Plan}
\usage{
JiraPlan(repos, datadir)
}
\arguments{
\item{repos}{Repository \code{data.table} object.}

\item{datadir}{Directory where the data is stored.}
}
\value{
A drake plan tibble.
}
\description{
Returns a drake plan for Jira.
}
