% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jira.R
\name{JiraTime}
\alias{JiraTime}
\title{Jira Time}
\usage{
JiraTime(table, name)
}
\arguments{
\item{table}{deta.frame containing raw time fields}

\item{name}{Name of the column in which to add the parsed time.}
}
\value{
The table with the time parsed as POSIXct added as a
  column.
}
\description{
Parses Jira raw time fields as POSIXct.
}
