% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlp.R
\name{LineType}
\alias{LineType}
\title{Line Type}
\usage{
LineType(type)
}
\arguments{
\item{type}{Type as a character string.}
}
\value{
Type as a factor with levels author, quote and generated.
}
\description{
Returns a factor corresponding to a specific line type
}
