% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{LocalTime}
\alias{LocalTime}
\title{Local time}
\usage{
LocalTime(time, tz)
}
\arguments{
\item{time}{The time object}

\item{tz}{The timezone formatted as +dddd or -dddd.}
}
\value{
The local time.
}
\description{
Converts a time to local time based on a timezone
}
