% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{LogPlan}
\alias{LogPlan}
\title{Log Plan}
\usage{
LogPlan(datadir)
}
\arguments{
\item{datadir}{Directory where the data is stored.}
}
\value{
A drake plan tibble.
}
\description{
Returns a drake plan for the different final log files
}
