% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{MapCommentsPlan}
\alias{MapCommentsPlan}
\title{Map Comments Plan}
\usage{
MapCommentsPlan(
  FUNC,
  target.name,
  input.subdir,
  input.subname,
  output.subdir,
  output.subname,
  repos,
  datadir,
  types = c("bugzilla", "jira")
)
}
\arguments{
\item{FUNC}{Function to apply on comments.}

\item{target.name}{Base name to use for the drake targets.}

\item{input.subdir}{Subdirectory of the input comments.}

\item{input.subname}{File sub name of the input comments.}

\item{output.subdir}{Subdirectory of the output comments.}

\item{output.subname}{File sub name of the output comments.}

\item{repos}{Repository \code{data.table} object.}

\item{datadir}{Directory where the data is stored.}

\item{types}{Types of issue repository to use.}
}
\value{
A drake plan tibble.
}
\description{
Returns a drakw plan for applying a function on issue comments.
}
