% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identities.R
\name{MergeByEmail}
\alias{MergeByEmail}
\title{Merge by email}
\usage{
MergeByEmail(people)
}
\arguments{
\item{people}{Identities \code{data.table}.}
}
\value{
Identities with the same merged.id for identities with the same email.
}
\description{
Merge identities by email.
}
