% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identities.R
\name{MergeByName}
\alias{MergeByName}
\title{Merge by name}
\usage{
MergeByName(people)
}
\arguments{
\item{people}{Identities \code{data.table}.}
}
\value{
Identities with the same merged.id for identities with the same name.
}
\description{
Merge identities by name.
}
