% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MozillaAapacheDataset.R
\docType{package}
\name{MozillaApacheDataset}
\alias{MozillaApacheDataset}
\title{MozillaAapacheDataset: data processing of 20-MAD dataset.}
\description{
Data processing functions for the 20 years of Mozilla and Apache
Development dataset.
}
