% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlp.R
\name{NLComments}
\alias{NLComments}
\title{NL Comments}
\usage{
NLComments(comments)
}
\arguments{
\item{comments}{\code{data.table} object with comments.}
}
\value{
\code{data.table} comments with columns source, product,
  issue.id, comments.id, paragraph.id, text and nchar.
}
\description{
Subset comments to only keep natural language comments from the
author and splits by paragraphs.
}
