% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlp.R
\name{NLoN}
\alias{NLoN}
\title{NLoN}
\usage{
NLoN(text, model)
}
\arguments{
\item{text}{The text.}

\item{model}{The NLoN model.}
}
\value{
\code{data.table} object with nlon boolean column and
  number of characters.
}
\description{
Use NLoN model to predict text.
}
