% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{NLPAggregatePlan}
\alias{NLPAggregatePlan}
\title{NLP Aggregate Plan}
\usage{
NLPAggregatePlan(repos, datadir, target, types = c("bugzilla", "jira"))
}
\arguments{
\item{repos}{Repository \code{data.table} object.}

\item{datadir}{Directory where the data is stored.}

\item{target}{Name of the target.}

\item{types}{Types of issue repository to use.}
}
\value{
A drake plan tibble.
}
\description{
Returns a drake plan for aggregating NLP plans.
}
