% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{NLPPlan}
\alias{NLPPlan}
\title{NLP Plan}
\usage{
NLPPlan(
  repos,
  datadir,
  senti4sd.chunk.limit = 1000,
  senti4sd.memory.limit = 0,
  types = c("bugzilla", "jira")
)
}
\arguments{
\item{repos}{Repository \code{data.table} object.}

\item{datadir}{Directory where the data is stored.}

\item{senti4sd.chunk.limit}{Maximum number of comments to process at the
same time with Senti4SD.}

\item{senti4sd.memory.limit}{Maximum amount of memory (in GB) to use for one
run of Senti4SD. Overrides \code{senti4sd.chunk.size} by setting
it to \code{500 * senti4sd.memory.limit}.}

\item{types}{Types of issue repository to use.}
}
\value{
A drake plan tibble.
}
\description{
Returns a drake plan for NLP of issue comments.
}
