% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{NLPPlanFunctions}
\alias{NLPPlanFunctions}
\title{NLP Plan Functions}
\usage{
NLPPlanFunctions(senti4sd.chunk.limit = 1000, senti4sd.memory.limit = 0)
}
\arguments{
\item{senti4sd.chunk.limit}{Maximum number of comments to process at the
same time with Senti4SD.}

\item{senti4sd.memory.limit}{Maximum amount of memory (in GB) to use for one
run of Senti4SD. Overrides \code{senti4sd.chunk.size} by setting
it to \code{500 * senti4sd.memory.limit}.}
}
\value{
A list of functions.
}
\description{
Functions for building comments map plans.
}
