% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{NLPPlans}
\alias{NLPPlans}
\title{NLP Plans}
\usage{
NLPPlans()
}
\value{
List of lists with names target, input.subdir,
  input.subname and output.subname.
}
\description{
Returns a list of map comments plans to build.
}
\seealso{
MapCommentsPlan
}
