% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identities.R
\name{ParseApacheEmail}
\alias{ParseApacheEmail}
\title{Parse Apache Email}
\usage{
ParseApacheEmail(email)
}
\arguments{
\item{email}{Email address.}
}
\value{
The email address with "dot" and "at" replaced by appropriate symbol.
}
\description{
Converts Apache email addresses obfuscated with "dot" and "at".
}
