% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{ParseApacheIssueKey}
\alias{ParseApacheIssueKey}
\title{Parse Apache Issue Key}
\usage{
ParseApacheIssueKey(log.message)
}
\arguments{
\item{log.message}{Vector of commit log messages.}
}
\value{
A vector where each element is a string of issue keys
  separated by spaces.
}
\description{
Parses issue keys from Apache commit log messages.
}
