% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{ParseCommit}
\alias{ParseCommit}
\title{Parse Commit}
\usage{
ParseCommit(commit, max.message.length = NA)
}
\arguments{
\item{commit}{The commit as a list (parsed from JSON).}

\item{max.message.length}{If not NA, shorten commit messages above
this length.}
}
\value{
A \code{data.table}.
}
\description{
Converts a commit given as a list to a \code{data.table}
}
