% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{ParseDiff}
\alias{ParseDiff}
\title{Parse Diff}
\usage{
ParseDiff(commit)
}
\arguments{
\item{commit}{The commit as a list (parsed from JSON).}
}
\value{
A \code{data.table} with columns hash, file, added, removed
  and action.
}
\description{
Converts a diff off a commit to \code{data.table}
}
