% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ParseExtension}
\alias{ParseExtension}
\title{Parse Extension}
\usage{
ParseExtension(x)
}
\arguments{
\item{x}{The filename.}
}
\value{
The file extension.
}
\description{
Parses the extension of a filename by taking into account multiple
successive extensions (such as tar.gz). Modified from
\code{tools::file_ext}.
}
