% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{ParseGitRepository}
\alias{ParseGitRepository}
\title{Parse Git Repository}
\usage{
ParseGitRepository(
  source,
  repo,
  file.in,
  log.out = NULL,
  diff.out = NULL,
  max.message.length = NA
)
}
\arguments{
\item{source}{Source of the repository (e.g. Apache or Mozilla).}

\item{repo}{Name of the Git repository.}

\item{file.in}{Input JSON file.}

\item{log.out}{Output Parquet file for the Git log.}

\item{diff.out}{Output Parquet file for the diff.}

\item{max.message.length}{If not NA, shorten commit messages above
this length.}
}
\value{
If output files are not NULL, returns the file names inside
  a \code{data.table} with columns source, repo, log and
  diff. Otherwise returns the \code{data.table} log and diff inside
  a list.
}
\description{
Parses a git repository JSON output as a log and diff.
}
