% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jira.R
\name{ParseJira}
\alias{ParseJira}
\title{Parse Jira}
\usage{
ParseJira(source, product, file.in, files.out = NULL)
}
\arguments{
\item{source}{Source of the issue tracker (Apache).}

\item{product}{Product tag of the issues.}

\item{file.in}{JSON file (possibly gzipped) containing the issues.}

\item{files.out}{List of output Parquet files for the issues,
comments and history.}
}
\description{
Parse a JSON log representing a set of Jira issues.
}
