% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{ParseMozillaIssueId}
\alias{ParseMozillaIssueId}
\title{Parse Mozilla Issue Id}
\usage{
ParseMozillaIssueId(log.message)
}
\arguments{
\item{log.message}{Vector of commit log messages.}
}
\value{
Vecotr of issue id integers (with possible NAs).
}
\description{
Parses issue ids from Mozilla commit log messages.
}
