% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ParseTZ}
\alias{ParseTZ}
\title{Parse time zone}
\usage{
ParseTZ(tz)
}
\arguments{
\item{tz}{The timezone formatted as +dddd or -dddd.}
}
\value{
The timezone as a number.
}
\description{
Parses a textual time zone as a number.
}
