% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlp.R
\name{ProcessApacheComments}
\alias{ProcessApacheComments}
\title{Process Apache Comments}
\usage{
ProcessApacheComments(comments)
}
\arguments{
\item{comments}{\code{data.table} object with comments.}
}
\value{
Processed Apache comments.
}
\description{
Processes Apache comments
}
