% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlp.R
\name{ProcessMozillaComments}
\alias{ProcessMozillaComments}
\title{Process Mozilla comments}
\usage{
ProcessMozillaComments(comments)
}
\arguments{
\item{comments}{\code{data.table} object with comments.}
}
\value{
Processed Mozilla comments.
}
\description{
Processes Mozilla comments.
}
