% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlp.R
\name{ProcessRawComments}
\alias{ProcessRawComments}
\title{Process Raw Comments}
\usage{
ProcessRawComments(comments, nlon.model = NULL, chunk.limit = 0)
}
\arguments{
\item{comments}{Raw comments to process.}

\item{nlon.model}{NLoN model.}

\item{chunk.limit}{Maximum number of comments to process with NLoN
in one chunk. 0 means no limit.}
}
\value{
The processed comments as a \code{data.table} with columns
  source, product, issue.id, comment.id, line.id, line.type, nchar
  and text.
}
\description{
Processes Apache and Mozilla raw comments and optionally runs NLoN.
}
\seealso{
ProcessApacheComments

ProcessMozillaComments
}
