% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlp.R
\name{ProcessText}
\alias{ProcessText}
\title{Process Text}
\usage{
ProcessText(comments, FUNC, chunk.limit = 0, ...)
}
\arguments{
\item{comments}{Comments \code{data.table} object.}

\item{FUNC}{Function to apply on comments.}

\item{chunk.limit}{Maximum number of comments to process in one
chunk. 0 means no limit.}

\item{...}{Addition parameters to pass to FUNC.}
}
\value{
Comments with additional columns returned from processing
  the comments with FUNC.
}
\description{
Processes text data with a function
}
