% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{ReadParquet}
\alias{ReadParquet}
\title{Read Parquet}
\usage{
ReadParquet(filename)
}
\arguments{
\item{filename}{Name of the Parquet file.}
}
\value{
The \code{data.table} object.
}
\description{
Reads a Parquet table from a file, converts it to a
\code{data.table} and replaces _ with . in the column names.
}
