% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{RemoveColumns}
\alias{RemoveColumns}
\title{Remove Columns}
\usage{
RemoveColumns(df, cols)
}
\arguments{
\item{df}{Data.frame or list.}

\item{cols}{Vector of columns.}
}
\description{
Removes columns from a data.frame or list.
}
