% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{RemoveNANames}
\alias{RemoveNANames}
\title{Remove NA Names}
\usage{
RemoveNANames(l)
}
\arguments{
\item{l}{List.}
}
\value{
The list without columns which names are NA.
}
\description{
Removes from a list columns which names are NA.
}
