% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{RemoveTextCols}
\alias{RemoveTextCols}
\title{Remove Text Columns}
\usage{
RemoveTextCols(df)
}
\arguments{
\item{df}{Data.frame object.}
}
\value{
The data.frame object without text and raw.text columns.
}
\description{
Removes "text" and "raw.text" columns from a data.frame.
}
