% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlp.R
\name{Senti4SD}
\alias{Senti4SD}
\title{Senti4SD}
\usage{
Senti4SD(comments, model, memory.limit = 0, chunk.limit = 1000)
}
\arguments{
\item{comments}{\code{data.table} object with comments.}

\item{model}{The LiblineaR model to use for Senti4SD.}

\item{memory.limit}{Maximum amount of memory (in GB) to use for one
run of Senti4SD. Overrides \code{senti4sd.chunk.size} by setting
it to \code{500 * senti4sd.memory.limit}.}

\item{chunk.limit}{Maximum number of comments to process at the
same time with Senti4SD.}
}
\value{
data.table as returned by \code{RSenti4SD::RunSenti4SD}.
}
\description{
Runs Senti4SD on comments.
}
