% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlp.R
\name{SentiStrength}
\alias{SentiStrength}
\title{SentiStrength}
\usage{
SentiStrength(comments, chunk.limit = 0, prefix = NULL)
}
\arguments{
\item{comments}{\code{data.table} object with comments.}

\item{chunk.limit}{Maximum number of comments to process at the
same time with SentiStrength.}

\item{prefix}{Prefix to add to the result's column names.}
}
\value{
comments with added columns valence.min, valence.max,
  arousal.min and arousal.max.
}
\description{
Runs SentiStrength on comments.
}
