% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{SimplifyGitlog}
\alias{SimplifyGitlog}
\title{Simplify Git Log}
\usage{
SimplifyGitlog(files, by.file.extension = FALSE)
}
\arguments{
\item{files}{List with the names of the Parquet files for the Git
log and diff.}

\item{by.file.extension}{If TRUE, number of added and removed lines
of code are summed by file extensions, otherwise summed by
commit.}
}
\value{
The commit log as a \code{data.table} with columns source,
  repo, hash, parents, author, author.time, author.tz, committer,
  commit.time, commit.tz, message, added and removed (and
  file.extension if \code{by.file.extension} is TRUE).
}
\description{
Simplifies a git log and mergres it with diff.
}
