% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identities.R
\name{SplitAuthor}
\alias{SplitAuthor}
\title{Split author}
\usage{
SplitAuthor(author.key)
}
\arguments{
\item{author.key}{The author fields.}
}
\value{
A list of splitted authors.
}
\description{
Split Git author fields based on various regex.
}
