% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlp.R
\name{SplitParagraphs}
\alias{SplitParagraphs}
\title{Split Paragraphs}
\usage{
SplitParagraphs(comments)
}
\arguments{
\item{comments}{\code{data.table} object with comments.}
}
\value{
\code{data.table} comments with columns source, product,
  issue.id, comments.id, paragraph.id, text and nchar.
}
\description{
Splits comments into paragraphs
}
