% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlp.R
\name{SplitSentences}
\alias{SplitSentences}
\title{Split Sentences}
\usage{
SplitSentences(comments)
}
\arguments{
\item{comments}{\code{data.table} object with comments.}
}
\value{
\code{data.table} comments with columns source, product,
  issue.id, comments.id, paragraph.id, sentence.id, text and nchar.
}
\description{
Split NL comments into sentences.
}
