% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identities.R
\name{StripNonAlpha}
\alias{StripNonAlpha}
\title{Strip Non Alpha}
\usage{
StripNonAlpha(s)
}
\arguments{
\item{s}{The character strings.}
}
\value{
The character strings stripped.
}
\description{
Strip non alpha numerical trailing characters from strings.
}
