% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{SubsetColumns}
\alias{SubsetColumns}
\title{Subset columns}
\usage{
SubsetColumns(table, cols)
}
\arguments{
\item{table}{\code{data.table} object.}

\item{cols}{Columns to subset.}
}
\value{
The subset of the table.
}
\description{
Takes a subset of columns from a \code{data.table} objects. Doesn't
throw error if columns are missing.
}
