% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{Timestamps}
\alias{Timestamps}
\title{Timestamps}
\usage{
Timestamps(
  commits.input,
  issues.input,
  comments.input,
  idmerging.input,
  tzhistory.input,
  output = NULL
)
}
\arguments{
\item{commits.input}{parquet file containing the commit log.}

\item{issues.input}{parquet file containing the issue log.}

\item{comments.input}{parquet file containing the issue comment
log.}

\item{idmerging.input}{Parquet file containing identity merging.}

\item{tzhistory.input}{Parquet file containing the commit timezone
history.}

\item{output}{Parquet file where to store the output.}
}
\value{
The output filename (if any) or the data.table object
  containing the timestamp log otherwise.
}
\description{
Returns list of timestamps for commits, issues and issue comments.
}
