% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{WriteParquet}
\alias{WriteParquet}
\title{Read Parquet}
\usage{
WriteParquet(df, filename, compression = "gzip")
}
\arguments{
\item{df}{The \code{data.table} object.}

\item{filename}{Name of the Parquet file.}

\item{compression}{Compression algorithm to use.}
}
\description{
Writes a \code{data.table} to a Parquet file after replacing . with
_ in the column names.
}
